alg911.catana-ai/requirements.txt
# Previous dependencies
supabase
pydantic
requests

# New dependencies for Protocol Polis
libp2p
pycryptodome
